\name{BioMart.list-class}
\docType{class}
\alias{BioMart.list-class}

\title{Class "BioMart.list"}
\description{Primitive Class - BioMart dataset info}
\section{Slots}{
	 \describe{
    \item{\code{s1}:}{object of class \code{"list"}: it  contains two string defining BioMart database and BioMart dataset  }
  }
}
\section{Methods}{
No methods defined with class "BioMart.list" in the signature.
}
\examples{
showClass("BioMart.list")
}
\keyword{classes}
