\name{Cluster.Dataframes}
\alias{Cluster.Dataframes}
\title{Summary Dataframes From a Cluster Analysis}
\description{ This function takes as input an object of class \code{Clusterama} and update the slot \code{summaries} of the object of class \code{Clusterama}.
}
\usage{
Cluster.Dataframes(clusterama = clusterama.object)
}
\arguments{
  \item{clusterama}{an object of class \code{Clusterama}}
}
\value{
}
\details{
}
\seealso{
}
\examples{
clusterama.obj <- Cluster.Dataframes(clusterama.obj)
}
