\name{CornaOut.df-class}
\docType{class}
\alias{CornaOut.df-class}

\title{Class "CornaOut.df"}
\description{Primitive Class - Output dataframe structure from \code{corna.test.fun} function}
\section{Slots}{
	 \describe{
    \item{\code{s1}:}{object of class \code{"data.frame"}  see documentation about \code{\link{corna.test.fun}}}
  }
}
\section{Methods}{
No methods defined with class "CornaOut.df" in the signature.
}
\examples{
showClass("CornaOut.df")
}
\keyword{classes}
