\name{Gene2Path.df-class}
\docType{class}
\alias{Gene2Path.df-class}

\title{Class "Gene2Path.df"}
\description{Primitive Class - crosslinks between Ensembl Genes and KEGG Pathways}
\section{Slots}{
	 \describe{
    \item{\code{s1}:}{object of class \code{"data.frame"}: each row contains info about Ensembl gene ID, KEGG Pathway ID and KEGG Pathway description}
  }
}
\section{Methods}{
No methods defined with class "Gene2Path.df" in the signature.
}
\examples{
showClass("Gene2Path.df")
}
\keyword{classes}
