\name{GO2Term.df-class}
\docType{class}
\alias{GO2Term.df-class}

\title{Class "GO2Term.df"}
\description{crosslinks between GO id and term}
\section{Slots}{
	 \describe{
    \item{\code{s1}:}{object of class \code{"data.frame"}: each row contains info about GO ID and GO term}
  }
}
\section{Methods}{
No methods defined with class "GO2Term.df" in the signature.
}
\examples{
showClass("GO2Term.df")
}
\keyword{classes}
