\name{MyMir.OverReps}
\alias{MyMir.OverReps}
\docType{data}
\title{Overall Test Output}
\description{
Example dataset after launching \code{OverallTest} function
}
\usage{data(MyMir.OverReps)}
\format{
object of class \code{"OverReps"}
}
\details{
This dataset is an example obtained from the following example:
}
\examples{
data(MyMir.UniPop)
MyMir.OverReps <- OverallTest(unipop.mir199.3p)
}
\keyword{datasets}