\name{Population.df-class}
\docType{class}
\alias{Population.df-class}
\alias{Universe.df-class}

\title{Classes "Population.df" and "Universe.df" }
\description{Primitive Class - dataframe containing "population" or "universe" subset}
\section{Slots}{
	 \describe{
    \item{\code{s1}:}{object of class \code{"data.frame"}: a "population" dataframe differs from a "universe" for two supplementary variables, \code{rank} and \code{description} }
  }
}
\examples{
showClass("Population.df")
}
\keyword{classes}
