\name{RefSeq2Ensembl.list-class}
\docType{class}
\alias{RefSeq2Ensembl.list-class}

\title{Class "RefSeq2Ensembl.list"}
\description{Primitive Class - crosslinks between RefSeq and Ensembl IDs}
\section{Slots}{
	 \describe{
    \item{\code{s1}:}{object of class \code{"list"} contains two dataframe for crosslinks between RefSeq and Ensembl Transcript or Ensembl Gene IDs respectively}
  }
}
\section{Methods}{
No methods defined with class "RefSeq2Ensembl.list" in the signature.
}
\examples{
showClass("RefSeq2Ensembl.list")
}
\keyword{classes}
