\name{Trans2GO.list-class}
\docType{class}
\alias{Trans2GO.list-class}

\title{Class "Trans2GO.list"}
\description{Primitive Class - crosslinks between Ensembl Transcripts and GO IDs}
\section{Slots}{
	 \describe{
    \item{\code{s1}:}{object of class \code{"list"} contains three dataframes defining links between Ensembl and GO biological process, GO cellular component or GO molecular function respectively}
  }
}
\examples{
showClass("Trans2GO.list")
}
\keyword{classes}
