\name{universe.mm9}
\alias{universe.mm9}
\docType{data}
\title{Mus Musculus whole genome dataset}
\description{
A non redundant dataframe containing Mus Musculus whole genome transcripts [assembly NCBI m37 (mm9)]
}
\usage{data(universe.mm9)}
\format{ A dataframe containing 4 variables: \cr
    \tabular{ll}{
    \code{$RefSeqID} \tab character vector of RefSeq identifiers \cr
    \code{$mm9} \tab character vector of RefSeq identifiers according to <kgXref> table of mymir DB \cr
    \code{$GeneID} \tab character vector of EntrezGene identifiers \cr
    \code{$GeneSymbol} \tab character vector of GeneSymbol according to MGI
    }
}
\keyword{datasets}
